backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.120.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.2
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.38.0
python-multipart==0.0.9 # Already needed for FastAPI, ensuring it's there
motor==3.3.2           # Asynchronous MongoDB driver for FastAPI
python-dotenv==1.0.1   # To manage environment variables (like your MongoDB URI)
bcrypt==4.1.3          # Secure password hashing library